feat: animation overhaul, currency fixes, budget persistence, status bar, and name field
Animations:
- Add flutter_animate, lottie, countup libraries
- Staggered section entrances on Home, Report, Plan, Subscriptions, People pages
- Numpad key stagger and amount entrance on Quick Add sheet
- Transaction list items stagger with fadeIn + slideX
- Lottie confetti on completed savings goals
- Lottie checkmark on onboarding completion screen
- Empty state widget supports Lottie with animated icon + text entrance
- 4 custom Lottie JSON assets (success_check, confetti, empty_wallet, onboarding_done)

Bug fixes:
- Fix onboarding data not persisting: invalidate providers after completion
- Wire up "Set Monthly Budget" CTA to navigate to Plan tab
- Fix monthly budget type mismatch (int vs double) across onboarding, settings, Firestore
- Replace 40+ hardcoded "$" with dynamic currency across plan, transactions, subscriptions, people, split flow pages
- Fix status bar icons invisible on dark headers (AnnotatedRegion per tab)
- Add "What should we call you?" name field to account creation
- Remove extra spacing below category in Quick Add sheet

+++ lib/pages/onboarding_v2/monthly_budget_screen.dart
@@ -110,7 +110,7 @@ class _MonthlyBudgetScreenState extends State<MonthlyBudgetScreen>
 
   Future<void> _onContinue() async {
     final prefs = await SharedPreferences.getInstance();
-    await prefs.setInt('monthly_budget', _selectedAmount);
+    await prefs.setDouble('monthly_budget', _selectedAmount.toDouble());
     if (mounted) await context.push('/onboarding/step4');
   }
